user template login stuctured with helpers
