+++ app/packs/src/fetchers/UsersFetcher.js
@@ -316,8 +316,6 @@ export default class UsersFetcher {
 
   static updateUserKetcher2Options(list) {
     const data = JSON.parse(list);
-    alert("hit!");
-
     const promise = fetch('/api/v1/profiles/editors/ketcher2-options', {
       credentials: 'same-origin',
       method: 'PUT',
